ADD: File options for media files: play, info, delete

+++ backend/core/files_handler.py
@@ -47,6 +47,10 @@ class FolderInfo(BaseModel):
         default="0 KB",
         description="Size of the file/folder in human-readable format (e.g. '10 KB').",
     )
+    path: str = Field(
+        ...,
+        description="Path of the file/folder.",
+    )
     files: list["FolderInfo"] = Field(
         default=[],
         description="List of FileInfo or FolderInfo objects representing files and folders \
@@ -95,6 +99,7 @@ class FilesHandler:
             type="file",
             name=unicodedata.normalize("NFKD", entry.name),
             size=FilesHandler._convert_file_size(info.st_size),
+            path=entry.path,
             created=dt.fromtimestamp(info.st_ctime).strftime("%Y-%m-%d %H:%M:%S"),
         )
 
@@ -138,6 +143,7 @@ class FilesHandler:
             name=unicodedata.normalize("NFKD", os.path.basename(folder_path)),
             files=dir_info,
             size=dir_size_str,
+            path=folder_path,
             created=dt.fromtimestamp(os.path.getctime(folder_path)).strftime(
                 "%Y-%m-%d %H:%M:%S"
             ),
@@ -318,6 +324,10 @@ class FilesHandler:
             file_path (str): Path to the file to delete.\n
         Returns:
             bool: True if the file is deleted successfully, False otherwise."""
+        # Make sure the path is at least 2 levels deep from the root
+        if file_path.count("/") < 3:
+            logger.error(f"Cannot delete top level file: {file_path}")
+            return False
         try:
             await aiofiles.os.remove(file_path)
             logger.debug(f"File deleted: {file_path}")
@@ -333,9 +343,18 @@ class FilesHandler:
     async def delete_folder(folder_path: str) -> bool:
         """Delete a folder from the filesystem.\n
         Args:
-            folder_path (str): Path to the folder to delete.\n
+            folder_path (str): Path to the folder to delete. \
+                Should be at least 3 folders deep from root. \n
         Returns:
             bool: True if the folder is deleted successfully, False otherwise."""
+        # Make sure we are not deleting the root folder or a top level folder
+        if folder_path == "/" or folder_path == "" or folder_path.count("/") < 3:
+            logger.error(f"Cannot delete root folder: {folder_path}")
+            return False
+        # Make sure the path is at least 3 levels deep from the root
+        if folder_path.count("/") < 3:
+            logger.error(f"Cannot delete top level folder: {folder_path}")
+            return False
         try:
             shutil.rmtree(folder_path)
             logger.debug(f"Folder deleted: {folder_path}")
@@ -418,3 +437,36 @@ class FilesHandler:
         msg += f"and {len(trailer_folders_inline)} (inline) trailers."
         logger.debug(msg)
         return trailer_folders_inline.union(trailer_folders)
+
+    @staticmethod
+    async def rename_file_fol(old_path: str, new_path: str) -> bool:
+        """Rename a file or folder.\n
+        Args:
+            old_path (str): Path of the file/folder to rename.
+            new_path (str): New path of the file/folder.
+        Returns:
+            bool: True if the file/folder is renamed successfully, False otherwise."""
+        try:
+            await aiofiles.os.rename(old_path, new_path)
+            logger.debug(f"File/Folder renamed: {old_path} -> {new_path}")
+            return True
+        except FileNotFoundError:
+            logger.error(f"File/Folder not found: {old_path}")
+            return False
+        except Exception as e:
+            logger.error(f"Failed to rename file/folder: {old_path}. Exception: {e}")
+            return False
+
+    @staticmethod
+    async def delete_file_fol(path: str) -> bool:
+        """Delete a file or folder from the filesystem.\n
+        Args:
+            path (str): Path to the file/folder to delete.\n
+        Returns:
+            bool: True if the file/folder is deleted successfully, False otherwise."""
+        # Check if the path is a file or folder
+        _is_dir = await aiofiles.os.path.isdir(path)
+        if _is_dir:
+            return await FilesHandler.delete_folder(path)
+        else:
+            return await FilesHandler.delete_file(path)
